Add MIT license to project
This commit adds the MIT license to the project, granting users the permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software. The MIT license is a permissive open-source license that allows for maximum freedom for users while protecting the original authors and contributors.